added manage asset button as footer
