+++ src/config/locales/en-US.json
@@ -97,6 +97,7 @@
     "engine_title":"Hive Engine Tokens",
     "engine_claim_btn":"Claim Rewards ({count})",
     "engine_select_tokens":"Select Tokens",
+    "manage_assets":"Add / Remove Assets",
     "ecency": {
       "title": "Points",
       "name":"Ecency Points",
