+++ src/screens/wallet/styles/manageAssets.styles.ts
@@ -3,11 +3,14 @@ import EStyleSheet from 'react-native-extended-stylesheet';
 
 export default EStyleSheet.create({
   container:{
-    marginBottom:12, marginTop:8
+    marginBottom:12, marginTop:8,
+    marginHorizontal: 16,
+    borderRadius:12,
+    borderWidth:4,
+    borderColor:'$primaryLightBackground'
   } as ViewStyle,
   engineBtnContainer: {
     alignItems: 'flex-end',
-    // 
     marginVertical: 8,
   } as TextStyle,
   engineBtnText: {
@@ -16,23 +19,24 @@ export default EStyleSheet.create({
   headerWrapper: {
     marginHorizontal: 16,
     flexDirection: 'row',
-    justifyContent: 'space-between',
+    justifyContent: 'center',
     alignItems: 'center'
   } as ViewStyle,
   title:{
-    fontSize:18,
+    fontSize:16,
   color: '$primaryBlack'
   } as TextStyle,
   rightIconWrapper: {
     alignSelf: 'center',
-    backgroundColor: '$primaryLightBackground',
     width: 40,
     height:40,
     borderRadius:20,
-    alignItems:'center',
-    justifyContent:'center',
+
   } as ViewStyle,
   rightIcon: {
     color: '$primaryBlack',
+    marginVertical: 12,
+    marginHorizontal: 8,
+
   } as ViewStyle,
 })
